+++ cpp/runtimes/libjoynr-runtime/websocket/LibJoynrWebSocketRuntime.cpp
@@ -94,6 +94,9 @@ LibJoynrWebSocketRuntime::~LibJoynrWebSocketRuntime()
     // WebSocketLibJoynrMessagingSkeleton
     websocket->registerReceiveCallback(nullptr);
     websocket->close();
+    // synchronously stop the underlying boost::asio::io_service
+    // this ensures all asynchronous operations are stopped now
+    // which allows a safe shutdown
     singleThreadIOService->stop();
 }